+++ src/Router/Router.php
@@ -79,11 +79,18 @@ class Router implements RouterInterface
                         }
                     }
 
-                    if (strlen($requestUrl['query'][$parameter])) {
-                        $parameters[$parameter] = $requestUrl['query'][$parameter];
-                    }
-                } elseif (strlen($default)) {
-                    $parameters[$parameter] = $default;
+                    $data = $requestUrl['query'][$parameter];
+                } else {
+                    $data = $default;
+                }
+
+                if (is_array($data) && current($data) === '') {
+                    // Empty array definition always contains an empty value, remove it and don't touch defined values.
+                    $data = array_slice($data, 1);
+                }
+
+                if ((is_array($data) && count($data)) || (is_string($data) && strlen($data))) {
+                    $parameters[$parameter] = $data;
                 }
             }
 
@@ -184,7 +191,7 @@ class Router implements RouterInterface
      *
      * @param string $url
      *
-     * @return array{path: array<int, string>, query: array<string, string>}
+     * @return array{path: array<int, string>, query: array<string, string|array<int, string>>}
      */
     private function parseUrl(string $url): array
     {