+++ src/Validator/Object/PropertiesValidator.php
@@ -6,6 +6,7 @@ namespace ExtendsSoftware\ExaPHP\Validator\Object;
 use ExtendsSoftware\ExaPHP\Validator\AbstractValidator;
 use ExtendsSoftware\ExaPHP\Validator\Exception\TemplateNotFound;
 use ExtendsSoftware\ExaPHP\Validator\Object\Properties\Property;
+use ExtendsSoftware\ExaPHP\Validator\Other\ProxyValidator;
 use ExtendsSoftware\ExaPHP\Validator\Result\Container\ContainerResult;
 use ExtendsSoftware\ExaPHP\Validator\Result\ResultInterface;
 use ExtendsSoftware\ExaPHP\Validator\Type\ObjectValidator;
@@ -28,22 +29,24 @@ class PropertiesValidator extends AbstractValidator
     public const PROPERTY_MISSING = 'propertyMissing';
 
     /**
+     * Properties.
+     *
+     * @var array<string, Property>
+     */
+    private array $properties = [];
+
+    /**
      * ObjectPropertiesValidator constructor.
      *
-     * @param mixed[]|null $properties
-     * @param bool|null    $strict
+     * @param array<string, ValidatorInterface>|null $properties
+     * @param bool|null                              $strict
      */
-    public function __construct(private ?array $properties = null, private ?bool $strict = null)
+    public function __construct(?array $properties = null, private ?bool $strict = null)
     {
-        $this->properties ??= [];
         $this->strict ??= true;
 
         foreach ($properties ?? [] as $property => $validator) {
-            if (is_array($validator)) {
-                [$validator, $optional] = $validator;
-            }
-
-            $this->addProperty($property, $validator, $optional ?? null);
+            $this->addProperty($property, $validator);
         }
     }
 
@@ -59,11 +62,10 @@ class PropertiesValidator extends AbstractValidator
         }
 
         $container = new ContainerResult();
-        /** @phpstan-ignore-next-line */
         foreach ($this->properties as $property) {
             $name = $property->getName();
             if (!property_exists($value, $name)) {
-                if (!$property->isOptional()) {
+                if (!$property->getValidator() instanceof ProxyValidator) {
                     $container->addResult(
                         $this->getInvalidResult(self::PROPERTY_MISSING, [
                             'property' => $name,
@@ -97,16 +99,12 @@ class PropertiesValidator extends AbstractValidator
      *
      * @param string             $property
      * @param ValidatorInterface $validator
-     * @param bool|null          $optional
      *
      * @return PropertiesValidator
      */
-    public function addProperty(
-        string             $property,
-        ValidatorInterface $validator,
-        bool               $optional = null
-    ): PropertiesValidator {
-        $this->properties[$property] = new Property($property, $validator, $optional);
+    public function addProperty(string $property, ValidatorInterface $validator): PropertiesValidator
+    {
+        $this->properties[$property] = new Property($property, $validator);
 
         return $this;
     }
@@ -136,7 +134,6 @@ class PropertiesValidator extends AbstractValidator
     private function checkStrictness(ContainerResult $container, mixed $object): void
     {
         foreach ($object as $property => $value) {
-            /** @phpstan-ignore-next-line */
             if (!array_key_exists($property, $this->properties)) {
                 $container->addResult(
                     $this->getInvalidResult(self::PROPERTY_NOT_ALLOWED, [